Move large DICOM file tests to GH1453
Add MemoryFailPoint to avoid test failing due to insufficient memory
