+++ Tests/FO-DICOM.Tests/DicomFileTest.cs
@@ -8,9 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.MemoryMappedFiles;
-using System.Threading;
 using System.Threading.Tasks;
-using FellowOakDicom.Imaging;
 using Xunit;
 
 namespace FellowOakDicom.Tests
@@ -452,156 +450,6 @@ namespace FellowOakDicom.Tests
             foreach (DicomItem item in clone.Dataset)
             {
                 Assert.Equal(file.Dataset.GetString(item.Tag), clone.Dataset.GetString(item.Tag));
-            }
-        }
-
-        [Theory]
-        [InlineData(FileReadOption.ReadLargeOnDemand)]
-        [InlineData(FileReadOption.ReadAll)]
-        [InlineData(FileReadOption.SkipLargeTags)]
-        public async Task SavingAndOpeningAVeryLargeDicomFile_ShouldWork(FileReadOption readOption)
-        {
-            var tempFileName = Path.GetTempFileName();
-            var disposables = new List<IDisposable>();
-            try
-            {
-                // Arrange
-                var pixelDataLength = 4_294_967_298; // 4 GB + 2 bytes, so the file will be larger than uint.MaxValue
-                var numberOfFrames = 2; // 2 GB + 1 byte, so each frame will be larger than int.MaxValue but smaller than uint.MaxValue
-                var frameLength = pixelDataLength / numberOfFrames;
-                var pixelData = new DicomOtherByteFragment(DicomTag.PixelData);
-                for (int i = 0; i < numberOfFrames; i++)
-                {
-                    var frameDataStream = new ConstantValueStream(frameLength, 255);
-                    pixelData.Fragments.Add(new StreamByteBuffer(frameDataStream, 0, frameLength));
-                    disposables.Add(frameDataStream);
-                }
-
-                var originalDicomFile = new DicomFile();
-                originalDicomFile.FileMetaInfo.TransferSyntax = DicomTransferSyntax.ExplicitVRLittleEndian;
-                originalDicomFile.Dataset.AddOrUpdate(DicomTag.SOPInstanceUID, DicomUIDGenerator.GenerateDerivedFromUUID());
-                originalDicomFile.Dataset.AddOrUpdate(DicomTag.BitsAllocated, (ushort) 8);
-                originalDicomFile.Dataset.AddOrUpdate(DicomTag.BitsStored, (ushort) 8);
-                originalDicomFile.Dataset.AddOrUpdate(DicomTag.NumberOfFrames, numberOfFrames);
-                originalDicomFile.Dataset.AddOrUpdate(pixelData);
-
-                // Act
-                await originalDicomFile.SaveAsync(tempFileName);
-
-                var openedDicomFileFromFileName = await DicomFile.OpenAsync(tempFileName, readOption);
-                using var fileStream = File.OpenRead(tempFileName);
-                var openedDicomFileFromStream = await DicomFile.OpenAsync(fileStream, readOption);
-
-                // Assert
-                if (readOption != FileReadOption.SkipLargeTags)
-                {
-                    foreach (var openedDicomFile in new[] { openedDicomFileFromFileName, openedDicomFileFromStream })
-                    {
-                        var openedPixelData =
-                            openedDicomFile.Dataset.GetDicomItem<DicomOtherByteFragment>(DicomTag.PixelData);
-                        Assert.Equal(numberOfFrames, openedPixelData.Fragments.Count);
-                        Assert.Equal(frameLength, openedPixelData.Fragments[0].Size);
-                        Assert.Equal(frameLength, openedPixelData.Fragments[numberOfFrames - 1].Size);
-                    }
-                }
-            }
-            finally
-            {
-                foreach (var disposable in disposables)
-                {
-                    disposable.Dispose();
-                }
-                File.Delete(tempFileName);
-            }
-        }
-
-        private class ConstantValueStream : Stream, IAsyncDisposable
-        {
-            public ConstantValueStream(long length, byte value)
-            {
-                Length = length;
-                Value = value;
-            }
-
-            public override bool CanRead => true;
-            public override bool CanSeek => true;
-            public override bool CanWrite => false;
-            public override long Length { get; }
-            public byte Value { get; }
-            public override long Position { get; set; }
-
-            public override void Flush()
-            {
-
-            }
-
-            public override int Read(byte[] buffer, int offset, int count)
-            {
-                var numberOfBytesToRead = (int) Math.Min(count, Length - Position);
-                if (numberOfBytesToRead == 0)
-                {
-                    return 0;
-                }
-
-                var max = Math.Min(offset + count, buffer.Length);
-                for (var i = offset; i < max; i++)
-                {
-                    buffer[i] = Value;
-                }
-                Position += numberOfBytesToRead;
-                return count;
-            }
-
-            public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-                var numberOfBytesToRead = (int) Math.Min(count, Length - Position);
-                if (numberOfBytesToRead == 0)
-                {
-                    return Task.FromResult(0);
-                }
-                var max = Math.Min(offset + count, buffer.Length);
-                for (var i = offset; i < max; i++)
-                {
-                    buffer[i] = Value;
-                }
-                Position += numberOfBytesToRead;
-                return Task.FromResult(count);
-            }
-
-            public override long Seek(long offset, SeekOrigin origin)
-            {
-                switch (origin)
-                {
-                    case SeekOrigin.Begin:
-                        Position = offset;
-                        break;
-                    case SeekOrigin.Current:
-                        Position += offset;
-                        break;
-                    case SeekOrigin.End:
-                        Position = Length - offset;
-                        break;
-                    default:
-                        throw new ArgumentOutOfRangeException(nameof(origin), origin, null);
-                }
-
-                return Position;
-            }
-
-            public override void SetLength(long value)
-            {
-                throw new NotSupportedException();
-            }
-
-            public override void Write(byte[] buffer, int offset, int count)
-            {
-                throw new NotSupportedException();
-            }
-
-            public ValueTask DisposeAsync()
-            {
-                return default;
             }
         }
